feat(back): fully migrated to sqlc

+++ backend/internal/services/jwt.go
@@ -11,11 +11,11 @@ import (
 var jwtKey = []byte(os.Getenv("JWT_SECRET_KEY"))
 
 type Claims struct {
-	UserID int64 `json:"user_id"`
+	UserID int32 `json:"user_id"`
 	jwt.StandardClaims
 }
 
-func GenerateJWT(userID int64) (string, error) {
+func GenerateJWT(userID int32) (string, error) {
 	expirationTime := time.Now().Add(240 * time.Hour)
 	claims := &Claims{
 		UserID: userID,
